+++ examples/cfd/external_aerodynamics/domino/src/train.py
@@ -203,10 +203,10 @@ def integral_loss_fn(
     output, target, area, normals, stream_velocity=None, padded_value=-10
 ):
     drag_loss = drag_loss_fn(
-        output, target, area, normals, stream_velocity, padded_value=-10
+        output, target, area, normals, stream_velocity=stream_velocity, padded_value=-10
     )
     lift_loss = lift_loss_fn(
-        output, target, area, normals, stream_velocity, padded_value=-10
+        output, target, area, normals, stream_velocity=stream_velocity, padded_value=-10
     )
     return lift_loss + drag_loss
 
@@ -298,7 +298,10 @@ def compute_loss_dict(
         surface_areas = batch_inputs["surface_areas"]
         surface_areas = torch.unsqueeze(surface_areas, -1)
         surface_normals = batch_inputs["surface_normals"]
-        stream_velocity = batch_inputs["stream_velocity"]
+
+        # Needs to be taken from the dataset
+        stream_velocity = batch_inputs["global_params_values"][:, 0, :]
+
         loss_surf = loss_fn_surface(
             prediction_surf,
             target_surf,
@@ -518,6 +521,16 @@ def main(cfg: DictConfig) -> None:
     else:
         num_surf_vars = None
 
+    num_global_features = 0
+    global_params_names = list(cfg.variables.global_parameters.keys())
+    for param in global_params_names:
+        if cfg.variables.global_parameters[param].type == "vector":
+            num_global_features += len(cfg.variables.global_parameters[param].reference)
+        elif cfg.variables.global_parameters[param].type == "scalar":
+            num_global_features += 1
+        else:
+            raise ValueError(f"Unknown global parameter type")
+
     vol_save_path = os.path.join(
         "outputs", cfg.project.name, "volume_scaling_factors.npy"
     )
@@ -580,6 +593,7 @@ def main(cfg: DictConfig) -> None:
         input_features=3,
         output_features_vol=num_vol_vars,
         output_features_surf=num_surf_vars,
+        global_features=num_global_features,
         model_parameters=cfg.model,
     ).to(dist.device)
     model = torch.compile(model, disable=True)  # TODO make this configurable