+++ physicsnemo/datapipes/cae/domino_sharded_datapipe.py
@@ -110,8 +110,8 @@ class ShardedDoMINODataPipe(DoMINODataPipe):
 
         # These keys are scalar-like and should never be sharded
         self.scalar_keys = [
-            "stream_velocity",
-            "air_density",
+            "global_params_values",
+            "global_params_reference",
             "surface_min_max",
             "volume_min_max",
             "length_scale",