+++ physicsnemo/models/domino/model.py
@@ -497,9 +497,9 @@ class ParameterModel(nn.Module):
     """
     Neural network module to encode simulation parameters.
 
-    This module encodes physical parameters such as inlet velocity and air density
-    into a learned latent representation that can be incorporated into the model's
-    prediction process.
+    This module encodes physical global parameters into a learned
+    latent representation that can be incorporated into the
+    model'sprediction process.
     """
 
     def __init__(self, input_features: int, model_parameters=None):
@@ -714,6 +714,7 @@ class DoMINO(nn.Module):
     >>> bsize = 1
     >>> nx, ny, nz = cfg.model.interp_res
     >>> num_neigh = 7
+    >>> global_features = 2
     >>> pos_normals_closest_vol = torch.randn(bsize, 100, 3).to(device)
     >>> pos_normals_com_vol = torch.randn(bsize, 100, 3).to(device)
     >>> pos_normals_com_surface = torch.randn(bsize, 100, 3).to(device)
@@ -732,8 +733,8 @@ class DoMINO(nn.Module):
     >>> volume_coordinates = torch.randn(bsize, 100, 3).to(device)
     >>> vol_grid_max_min = torch.randn(bsize, 2, 3).to(device)
     >>> surf_grid_max_min = torch.randn(bsize, 2, 3).to(device)
-    >>> stream_velocity = torch.randn(bsize, 1).to(device)
-    >>> air_density = torch.randn(bsize, 1).to(device)
+    >>> global_params_values = torch.randn(bsize, global_features, 1).to(device)
+    >>> global_params_reference = torch.randn(bsize, global_features, 1).to(device)
     >>> input_dict = {
     ...            "pos_volume_closest": pos_normals_closest_vol,
     ...            "pos_volume_center_of_mass": pos_normals_com_vol,
@@ -753,8 +754,8 @@ class DoMINO(nn.Module):
     ...            "volume_mesh_centers": volume_coordinates,
     ...            "volume_min_max": vol_grid_max_min,
     ...            "surface_min_max": surf_grid_max_min,
-    ...            "stream_velocity": stream_velocity,
-    ...            "air_density": air_density,
+    ...            "global_params_reference": global_params_values,
+    ...            "global_params_values": global_params_reference,
     ...        }
     >>> output = model(input_dict)
     >>> print(f"{output[0].shape}, {output[1].shape}")
@@ -766,6 +767,7 @@ class DoMINO(nn.Module):
         input_features: int,
         output_features_vol: int | None = None,
         output_features_surf: int | None = None,
+        global_features: int = 2,
         model_parameters=None,
     ):
         """
@@ -784,6 +786,7 @@ class DoMINO(nn.Module):
         self.input_features = input_features
         self.output_features_vol = output_features_vol
         self.output_features_surf = output_features_surf
+        self.global_features = global_features
 
         if self.output_features_vol is None and self.output_features_surf is None:
             raise ValueError(
@@ -807,7 +810,6 @@ class DoMINO(nn.Module):
         self.use_surface_normals = model_parameters.use_surface_normals
         self.use_surface_area = model_parameters.use_surface_area
         self.encode_parameters = model_parameters.encode_parameters
-        self.param_scaling_factors = model_parameters.parameter_model.scaling_params
         self.geo_encoding_type = model_parameters.geometry_encoding_type
 
         if self.use_surface_normals:
@@ -822,7 +824,8 @@ class DoMINO(nn.Module):
             # Defining the parameter model
             base_layer_p = model_parameters.parameter_model.base_layer
             self.parameter_model = ParameterModel(
-                input_features=2, model_parameters=model_parameters.parameter_model
+                input_features=self.global_features,
+                model_parameters=model_parameters.parameter_model,
             )
         else:
             base_layer_p = 0
@@ -1109,8 +1112,8 @@ class DoMINO(nn.Module):
         surface_neighbors_normals,
         surface_areas,
         surface_neighbors_areas,
-        inlet_velocity,
-        air_density,
+        global_params_values,
+        global_params_reference,
     ):
         """Function to approximate solution given the neighborhood information"""
         num_variables = self.num_variables_surf
@@ -1119,24 +1122,19 @@ class DoMINO(nn.Module):
         num_sample_points = surface_mesh_neighbors.shape[2] + 1
 
         if self.encode_parameters:
-            inlet_velocity = torch.unsqueeze(inlet_velocity, 1)
-            inlet_velocity = inlet_velocity.expand(
-                inlet_velocity.shape[0],
-                surface_mesh_centers.shape[1],
-                inlet_velocity.shape[2],
-            )
-            inlet_velocity = inlet_velocity / self.param_scaling_factors[0]
-
-            air_density = torch.unsqueeze(air_density, 1)
-            air_density = air_density.expand(
-                air_density.shape[0],
-                surface_mesh_centers.shape[1],
-                air_density.shape[2],
-            )
-            air_density = air_density / self.param_scaling_factors[1]
-
-            params = torch.cat((inlet_velocity, air_density), dim=-1)
-            param_encoding = self.parameter_model(params)
+            processed_parameters = []
+            for k in range(global_params_values.shape[1]):
+                param = torch.unsqueeze(global_params_values[:, k, :], 1)
+                ref = torch.unsqueeze(global_params_reference[:, k, :], 1)
+                param = param.expand(
+                    param.shape[0],
+                    surface_mesh_centers.shape[1],
+                    param.shape[2],
+                )
+                param = param / ref
+                processed_parameters.append(param)
+            processed_parameters = torch.cat(processed_parameters, axis=-1)
+            param_encoding = self.parameter_model(processed_parameters)
 
         if self.use_surface_normals:
             if not self.use_surface_area:
@@ -1271,8 +1269,8 @@ class DoMINO(nn.Module):
         volume_mesh_centers,
         encoding_g,
         encoding_node,
-        inlet_velocity,
-        air_density,
+        global_params_values,
+        global_params_reference,
         eval_mode,
         num_sample_points=20,
         noise_intensity=50,
@@ -1288,22 +1286,19 @@ class DoMINO(nn.Module):
             agg_model = self.agg_model_surf
 
         if self.encode_parameters:
-            inlet_velocity = torch.unsqueeze(inlet_velocity, 1)
-            inlet_velocity = inlet_velocity.expand(
-                inlet_velocity.shape[0],
-                volume_mesh_centers.shape[1],
-                inlet_velocity.shape[2],
-            )
-            inlet_velocity = inlet_velocity / self.param_scaling_factors[0]
-
-            air_density = torch.unsqueeze(air_density, 1)
-            air_density = air_density.expand(
-                air_density.shape[0], volume_mesh_centers.shape[1], air_density.shape[2]
-            )
-            air_density = air_density / self.param_scaling_factors[1]
-
-            params = torch.cat((inlet_velocity, air_density), dim=-1)
-            param_encoding = self.parameter_model(params)
+            processed_parameters = []
+            for k in range(global_params_values.shape[1]):
+                param = torch.unsqueeze(global_params_values[:, k, :], 1)
+                ref = torch.unsqueeze(global_params_reference[:, k, :], 1)
+                param = param.expand(
+                    param.shape[0],
+                    volume_mesh_centers.shape[1],
+                    param.shape[2],
+                )
+                param = param / ref
+                processed_parameters.append(param)
+            processed_parameters = torch.cat(processed_parameters, axis=-1)
+            param_encoding = self.parameter_model(processed_parameters)
 
         if self.solution_calculation_mode == "one-loop":
 
@@ -1450,8 +1445,8 @@ class DoMINO(nn.Module):
         surf_min = data_dict["surface_min_max"][:, 0]
 
         # Parameters
-        stream_velocity = data_dict["stream_velocity"]
-        air_density = data_dict["air_density"]
+        global_params_values = data_dict["global_params_values"]
+        global_params_reference = data_dict["global_params_reference"]
 
         if self.output_features_vol is not None:
             # Represent geometry on computational grid
@@ -1526,8 +1521,8 @@ class DoMINO(nn.Module):
                 volume_mesh_centers,
                 encoding_g_vol,
                 encoding_node_vol,
-                stream_velocity,
-                air_density,
+                global_params_values,
+                global_params_reference,
                 eval_mode="volume",
             )
         else:
@@ -1556,8 +1551,8 @@ class DoMINO(nn.Module):
                     surface_mesh_centers,
                     encoding_g_surf,
                     encoding_node_surf,
-                    stream_velocity,
-                    air_density,
+                    global_params_values,
+                    global_params_reference,
                     eval_mode="surface",
                     num_sample_points=1,
                     noise_intensity=500,
@@ -1572,8 +1567,8 @@ class DoMINO(nn.Module):
                     surface_neighbors_normals,
                     surface_areas,
                     surface_neighbors_areas,
-                    stream_velocity,
-                    air_density,
+                    global_params_values,
+                    global_params_reference,
                 )
         else:
             output_surf = None